requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-comments-dab==2.8.0
django-cors-headers==3.13.0
django-countries==7.5.1
django-crispy-forms==2.0
django-crontab==0.7.1
django-crum==0.7.9
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.0.2
django-jazzmin==2.6.0
django-mailer==2.2
django-model-utils==4.3.1
django-notifications==0.1.dev0
django-notifications-hq==1.7.0
django-rest-framework==0.1.0
django-select2==8.1.0
django-simpleui==2023.2.10
django-site-broadcasts==0.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
install==1.3.5
jsonfield==3.1.0
kombu==5.2.4
lockfile==0.12.2
MarkupPy==1.14
names==0.3.0
notifications==0.3.2
odfpy==1.4.1
openpyxl==3.1.1
Pillow==8.1.0
prompt-toolkit==3.0.37
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tablib==3.3.0
tinycss2==1.1.1
twilio==7.16.4
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
